Fixing the regex match in update-notifier.js
Previous regex would match a URL string with a hashtag considering it to be a page fragment, hence with an assumption that it can only occur as the last part of the URL.
However, in single page applications we sometimes have a sign that acts as a delimiter between a base URL and routes leading to other locations of the single page application.
In AngularJS for instance the # is used as the delimiter.

Fixing the regex to match a page fragment only for # that has no slash sign after it. Any other hashtag found between slash signs will be considered as part of the hostString.